smart-api-composer-full/requirements.txt
google-genai
fastapi
uvicorn
pydantic
pyyaml
httpx
sqlalchemy
aiosqlite
python-dotenv
